split filters and markdown settings into files
Cleans up the config file by splitting off functionality to the _11ty directory.
